refactor: split token helper to token decoder and token publisher
