+++ src/main/kotlin/com/kroffle/knitting/usecase/auth/AuthService.kt
@@ -3,19 +3,19 @@ package com.kroffle.knitting.usecase.auth
 import java.net.URI
 import java.util.UUID
 
-class AuthService(private val oAuthHelper: GoogleOAuthHelper, private val tokenHelper: TokenHelper) {
+class AuthService(private val oAuthHelper: GoogleOAuthHelper, private val tokenPublisher: TokenPublisher) {
     fun getAuthorizationUri(): URI = oAuthHelper.getAuthorizationUri()
 
     fun authorize(): String {
         // TODO Implement
-        return tokenHelper.publish(UUID.randomUUID())
+        return tokenPublisher.publish(UUID.randomUUID())
     }
 
     interface GoogleOAuthHelper {
         fun getAuthorizationUri(): URI
     }
 
-    interface TokenHelper {
+    interface TokenPublisher {
         fun publish(id: UUID): String
     }
 }
